+++ kernel/timer.c
@@ -40,7 +40,7 @@ void timer_install()
 /* This will continuously loop until the given time has been reached */
 void timer_wait(int ticks)
 {
-    unsigned long eticks;
+    int eticks;
 
     eticks = timer_ticks + ticks;
     while(timer_ticks < eticks);
